fix(core): keep React off the main entry point, and refresh the docs
The React boundary was not actually held. kubectlMapper imported kindToResource
from access.ts, whose first line imports React, so `import { toKubectl } from
"@srelens/core"` pulled React in transitively through the entry point that
advertises it does not.

The guard missed it because it looked for a direct `from "react"` in each
module and exempted the two hook modules by name. It now walks the import graph
from the barrel and fails if React is reachable at any depth, printing the chain
rather than just the culprit -- the edge worth breaking is usually the one
before the module that imports React. Verified by restoring the old import and
watching it fail.

kindToResource is a pure Kind -> {group, resource} lookup mirroring the
backend's gvk_for, so it belongs with the other kind tables in kinds.ts rather
than in a module of hooks. Its test moves with it.

docs/DEVELOPMENT.md pointed contributors at apps/desktop/src/lib and
apps/desktop/src/transport for the data layer, the transport shim and the
capability workflow, all of which are now @srelens/core. Its coverage section
also claimed 80% lines ratcheting toward 85, which was already out of date
before this change: the floors are 85/80/76, they now live in the root config
because they are measured across both packages, and `pnpm -r test` does not
enforce them.

Raised in review on #311.
